Passa os parametros do produto na migration

+++ database/migrations/2025_02_23_201922_add_fields_to_users_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->date('birth_date')->required();
             $table->string('cpf')->unique()->required();
             $table->string('photo')->nullable();
-            $table->float('balance')->default(0)->required();
+            $table->decimal('balance')->default(0)->required();
         });
     }
